refactor: move load_module_object to IOService as static method
